requirements.txt
urllib3<2
openai
langchain
chromadb
pypdf
tiktoken
youtube-transcript-api
pytube
python-dotenv
streamlit_chat
streamlit

